+++ src/containerapp/HISTORY.rst
@@ -8,6 +8,8 @@ upcoming
 * [Breaking Change] 'az containerapp service': deprecate command from Azure CLI version 2.56.1
 * 'az containerapp add-on' : support for add-on weaviate create and delete commands
 * Upgrade api-version to 2023-11-02-preview
+* 'az containerapp create/update/up': support --build-env-vars to set environment variables for build
+* 'az containerapp create/update': support --max-inactive-revisions
 * 'az containerapp env dapr-component resiliency': Add support for Dapr Component Resiliency Circuit Breakers
 
 0.3.46